css/css-grid/grid-model/grid-container-margin-border-padding-scrollbar-001.html

<!DOCTYPE html>
<html>
<head>
<title>CSS Grid: grid container's size includes border, padding and scrollbar.</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com"/>
<link rel="help" href="https://drafts.csswg.org/css-grid/#intrinsic-sizes"/>
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#overflow"/>
<meta name="assert" content="This test checks that grid container size includes border, padding and scrollbar; ignoring margin as expected."/>
<link rel="issue" href="https://crbug.com/532032"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="/css/support/width-keyword-classes.css" rel="stylesheet">
<link href="/css/support/height-keyword-classes.css" rel="stylesheet">
<style>
.margin {
    margin: 10px;
}

.border {
    border: 5px solid black;
}

.padding {
    padding: 20px;
}

.scroll {
    overflow: scroll;
}

.item {
    width: 100px;
    height: 100px;
    background: lime;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
</head>
<body onload="checkLayout('.grid');">
    <div class="grid min-content" data-expected-width="100" data-expected-height="100">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin" data-expected-width="100" data-expected-height="100">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content border" data-expected-width="110" data-expected-height="110">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content padding" data-expected-width="140" data-expected-height="140">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content scroll" data-expected-width="115" data-expected-height="115">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin border" data-expected-width="110" data-expected-height="110">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin padding" data-expected-width="140" data-expected-height="140">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin scroll" data-expected-width="115" data-expected-height="115">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content border padding" data-expected-width="150" data-expected-height="150">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content border scroll" data-expected-width="125" data-expected-height="125">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content padding scroll" data-expected-width="155" data-expected-height="155">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin border padding" data-expected-width="150" data-expected-height="150">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin border scroll" data-expected-width="125" data-expected-height="125">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin padding scroll" data-expected-width="155" data-expected-height="155">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content border padding scroll" data-expected-width="165" data-expected-height="165">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <div class="grid min-content margin border padding scroll" data-expected-width="165" data-expected-height="165">
        <div class="item" data-expected-width="100" data-expected-height="100"></div>
    </div>

    <!-- This div is only for measuring scrollbar size -->
    <div id="measure" style="height: 100px; width: 100px; display: inline-block; overflow: auto;">
        <div style="min-height: 300px;"></div>
    </div>

<script>
  var measure = document.getElementById('measure');
  var scrollbarSize = measure.offsetWidth - measure.clientWidth;

  // Here are the expected widths of all elements that have "scroll"
  // class. The things that contribute to the expected widths are:
  //
  // padding{Left,Right}, border{Left,Right}, element width and its scrollbarWidth.
  //
  // The attributes are dynamically set to the elements so that we can
  // ensure the scrollbar width is calculated in an engine-agnostic way.
  var expectedWidth = [100, 100, 110, 140, 110, 140, 150, 150];
  var elements = document.getElementsByClassName("scroll");
  for (var i = 0; i < elements.length; i++) {
    elements[i].setAttribute("data-expected-width", expectedWidth[i] + scrollbarSize);
  }
</script>

</body>
</html>
